+++ src/melee/pl/player.h
@@ -240,7 +240,7 @@ bool Player_GetMoreFlagsBit5(s32 slot);
 void Player_SetMoreFlagsBit5(s32 slot, u8 bit5);
 bool Player_GetMoreFlagsBit6(s32 slot);
 void Player_SetMoreFlagsBit6(s32 slot, u8 bit6);
-u8 Player_GetFlagsAEBit0(s32 slot);
+int Player_GetFlagsAEBit0(int slot);
 s32 Player_GetRemainingHPByIndex(s32 slot, s32 index);
 s32 Player_GetFalls(s32 slot);
 s32 Player_GetFallsByIndex(s32 slot, s32 index);
@@ -264,7 +264,7 @@ void Player_UnsetFlagsBit1(int slot);
 s32 Player_GetFlagsBit3(s32 slot);
 void Player_SetFlagsBit3(s32 slot, u8 bit3);
 int Player_GetFlagsBit4(int slot);
-u8 Player_GetFlagsBit5(s32 slot); ///< player is metal
+u8 Player_GetFlagsBit5(s32 slot);                ///< player is metal
 void Player_SetFlagsBit5(s32 slot, u8 is_metal); ///< player is metal
 u8 Player_GetFlagsBit6(s32 slot);
 void Player_SetFlagsBit6(s32 slot, u8 bit6);
